+++ src/zones/Pond.sol
@@ -27,6 +27,6 @@ contract Pond is Zone {
         override
         returns (uint256)
     {
-        return (hopper.strength * hopper.level * 10e8) / (10 * 100);
+        return (uint256(hopper.strength) * uint256(hopper.level) * 10e8) / (10 * 100);
     }
 }
